refactor:Config 상수 분리 dquote> dquote> 이전 : Config 파일에서 모두 관리 dquote> 변경 : Lotto 객체에서 LOTTO_LENGTH, LOTTO_PRICE, LottoNumber 객체에서 랜덤 최솟값, 최댓값 관리
