+++ src/main/kotlin/lotto/domain/LottoNumber.kt
@@ -1,8 +1,9 @@
 package lotto.domain
 
-import lotto.global.Config.MAX_RANDOM_NUM
-import lotto.global.Config.MIN_RANDOM_NUM
 import lotto.global.Message
+
+const val MAX_RANDOM_NUM = 45
+const val MIN_RANDOM_NUM = 1
 
 class LottoNumber private constructor(
     val value: Int,
