--- src/main/kotlin/lotto/global/Config.kt
@@ -1,9 +1,0 @@
-package lotto.global
-
-object Config {
-    const val MAX_LOTTO_LENGTH = 6
-    const val MAX_RANDOM_NUM = 45
-    const val MIN_RANDOM_NUM = 1
-    const val LOTTO_PRICE = 1000
-    const val RANDOM_SEED = 213498
-}
